Added WorldLibrary.
This implementation of WorldLibrary is different from the one in 0.4 server. This fetches only the assets that are in scene. It doesn't lag the whole Asset server by fetching every asset from the database. Because the way of the implementation, WorldLibrary can't fetch Flash animations because there is no reference to them in scene.

This module is tested with 3D models, Particle scripts, sounds and textures.
Untested: 3D animations and material scripts
Known not to work: Flash animations